backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
